chore: move analytics into its own module
