+++ web/src/analytics/mod.rs
@@ -1,0 +1,3 @@
+pub mod record;
+pub mod routes;
+pub mod visits;
