+++ web/src/routes/mod.rs
@@ -1,4 +1,3 @@
-pub mod analytics;
 mod context;
 pub mod index;
 pub mod portfolio;
