+++ web/src/state.rs
@@ -1,5 +1,5 @@
+use crate::analytics::{record::UniqueId, visits::Visits};
 use crate::prefetch::PrefetchedCollection;
-use crate::routes::analytics::{UniqueId, Visits};
 use std::{
     collections::HashSet,
     sync::{Arc, Mutex},
@@ -11,7 +11,6 @@ pub struct AppState {
     pub api_host: String,
     pub prefetched: PrefetchedCollection,
     pub ua_parser: UserAgentParser,
-    // pub visitor_etags: Arc<Mutex<HashMap<UniqueId, String>>>,
     pub unique_sessions: Arc<Mutex<HashSet<UniqueId>>>,
     pub visits: Arc<Mutex<Visits>>,
 }
